Move DataHandlers to a sub-package to differentiate MessageHandler and the DataHandlers
