+++ fe2-android/app/src/main/java/com/github/dedis/popstellar/utility/handler/MessageHandler.java
@@ -21,6 +21,7 @@ import com.github.dedis.popstellar.repository.LAOState;
 import com.github.dedis.popstellar.utility.error.DataHandlingException;
 import com.github.dedis.popstellar.utility.error.UnknownDataActionException;
 import com.github.dedis.popstellar.utility.error.UnknownDataObjectException;
+import com.github.dedis.popstellar.utility.handler.data.HandlerContext;
 
 import java.util.Collections;
 import java.util.List;
