+++ fe2-android/app/src/main/java/com/github/dedis/popstellar/utility/handler/data/ConsensusHandler.java
@@ -1,4 +1,4 @@
-package com.github.dedis.popstellar.utility.handler;
+package com.github.dedis.popstellar.utility.handler.data;
 
 import android.util.Log;
