+++ fe2-android/app/src/main/java/com/github/dedis/popstellar/utility/handler/data/DataHandler.java
@@ -1,4 +1,4 @@
-package com.github.dedis.popstellar.utility.handler;
+package com.github.dedis.popstellar.utility.handler.data;
 
 import com.github.dedis.popstellar.model.network.method.message.data.Data;
 import com.github.dedis.popstellar.utility.error.DataHandlingException;
